refactor: rename files to lower case
