--- src/components/LocaleSwitcher/index.ts
@@ -1,1 +1,0 @@
-export { LocaleSwitcher } from './LocaleSwitcher';
