--- src/components/Select/index.ts
@@ -1,1 +1,0 @@
-export { Select } from './Select';
