+++ src/components/icons/index.ts
@@ -1,2 +1,2 @@
-export { Icon } from './Icon';
+export { Icon } from './icon';
 export type { PhosphorIcons } from './phosphorIcons';
