+++ src/components/locale-switcher/index.ts
@@ -1,0 +1,1 @@
+export { LocaleSwitcher } from './locale-switcher';
