+++ src/components/locale-switcher/locale-switcher.tsx
