+++ src/components/translated/index.ts
@@ -1,0 +1,1 @@
+export { Translated } from './translated';
