+++ src/components/translated/translated.tsx
